Automize extracting trackpoints

+++ extract.sh
@@ -1,18 +1,28 @@
 #! /bin/bash
 set -e
 
-INPUT_FILE="${1}"
-filename="${INPUT_FILE%.*}"
-BIN_FILE="${filename}.bin"
-# GPX_FILE="${filename}.gpx"
-JSON_FILE="${filename}.json"
+DIR="/run/media/fabian/gopro/Gopro"
+
+pushd "${DIR}"
+
+for i in *.MP4; do
+    filename="${i%.*}"
+    BIN_FILE="${filename}.bin"
+    JSON_FILE="${filename}.json"
+
+    if [ ! -f ${JSON_FILE} ]; then
+        echo "${i} not yet processed"
+        ffmpeg -y -i ${i} -codec copy -map 0:2 -f rawvideo ${BIN_FILE}
+        docker run --rm -it --volume ${DIR}:/work/ maestroalubia/gopro gopro2json -i "/work/${BIN_FILE}" -o "/work/${JSON_FILE}"
+        rm -f ${BIN_FILE}
+    fi
+done
+
+popd
+
 # CSV_FILE="${filename}.csv"
-
-DIR=$(pwd)
-
-ffmpeg -y -i ${1} -codec copy -map 0:2 -f rawvideo ${BIN_FILE}
+# GPX_FILE="${filename}.gpx"
 # docker run --rm -it --volume ${DIR}:/work/ maestroalubia/gopro gopro2gpx -i "/work/${BIN_FILE}" -o "/work/${GPX_FILE}" -f 0 -a 100000
-docker run --rm -it --volume ${DIR}:/work/ maestroalubia/gopro gopro2json -i "/work/${BIN_FILE}" -o "/work/${JSON_FILE}"
+# docker run --rm -it --volume ${DIR}:/work/ maestroalubia/gopro gopro2json -i "/work/${BIN_FILE}" -o "/work/${JSON_FILE}"
 # docker run --rm -it --volume ${DIR}:/work/ maestroalubia/gopro gpmd2csv -i "/work/${BIN_FILE}" -o "/work/${CSV_FILE}"
 
-rm -f ${BIN_FILE}
